ugly warning fix for 64b - refs #33

+++ src/server/sv_client.c
@@ -741,7 +741,7 @@ void SV_SendClientGameState(client_t *client)
 	MSG_WriteLong(&msg, sv.checksumFeed);
 
 	// NERVE - SMF - debug info
-	Com_DPrintf("Sending %i bytes in gamestate to client: %i\n", msg.cursize, client - svs.clients);
+	Com_DPrintf("Sending %i bytes in gamestate to client: %i\n", msg.cursize, (int) (client - svs.clients));
 
 	// deliver this to the client
 	SV_SendMessageToClient(&msg, client);
@@ -1051,7 +1051,7 @@ void SV_WriteDownloadToClient(client_t *cl, msg_t *msg)
 		if (cl->downloadnotify & DLNOTIFY_BEGIN)
 		{
 			cl->downloadnotify &= ~DLNOTIFY_BEGIN;
-			Com_Printf("clientDownload: %d : beginning \"%s\"\n", cl - svs.clients, cl->downloadName);
+			Com_Printf("clientDownload: %d : beginning \"%s\"\n", (int) (cl - svs.clients), cl->downloadName);
 		}
 
 		idPack = FS_idPak(cl->downloadName, BASEGAME);
@@ -1062,12 +1062,12 @@ void SV_WriteDownloadToClient(client_t *cl, msg_t *msg)
 			// cannot auto-download file
 			if (idPack)
 			{
-				Com_Printf("clientDownload: %d : \"%s\" cannot download id pk3 files\n", cl - svs.clients, cl->downloadName);
+				Com_Printf("clientDownload: %d : \"%s\" cannot download id pk3 files\n", (int) (cl - svs.clients), cl->downloadName);
 				Com_sprintf(errorMessage, sizeof(errorMessage), "Cannot autodownload official pk3 file \"%s\"", cl->downloadName);
 			}
 			else
 			{
-				Com_Printf("clientDownload: %d : \"%s\" download disabled", cl - svs.clients, cl->downloadName);
+				Com_Printf("clientDownload: %d : \"%s\" download disabled", (int) (cl - svs.clients), cl->downloadName);
 				if (sv_pure->integer)
 				{
 					Com_sprintf(errorMessage, sizeof(errorMessage), "Could not download \"%s\" because autodownloading is disabled on the server.\n\n"
@@ -1159,7 +1159,7 @@ void SV_WriteDownloadToClient(client_t *cl, msg_t *msg)
 		cl->downloadSize = FS_SV_FOpenFileRead(cl->downloadName, &cl->download);
 		if (cl->downloadSize <= 0)
 		{
-			Com_Printf("clientDownload: %d : \"%s\" file not found on server\n", cl - svs.clients, cl->downloadName);
+			Com_Printf("clientDownload: %d : \"%s\" file not found on server\n", (int) (cl - svs.clients), cl->downloadName);
 			Com_sprintf(errorMessage, sizeof(errorMessage), "File \"%s\" not found on server for autodownloading.\n", cl->downloadName);
 			SV_BadDownload(cl, msg);
 			MSG_WriteString(msg, errorMessage);   // (could SV_DropClient isntead?)
@@ -2102,7 +2102,7 @@ void SV_ExecuteClientMessage(client_t *cl, msg_t *msg)
 
 	if (c != clc_EOF)
 	{
-		Com_Printf("WARNING: bad command byte for client %i\n", cl - svs.clients);
+		Com_Printf("WARNING: bad command byte for client %i\n", (int) (cl - svs.clients));
 	}
 
 	SV_ParseBinaryMessage(cl, msg);
